added main class
