+++ v2/elasticsearch/src/main/java/com/google/cloud/teleport/v2/elasticsearch/transforms/PubSubMessageToJsonDocument.java
@@ -1,0 +1,2 @@
+package com.google.cloud.teleport.v2.elasticsearch.templates;public class PubSubMessageToJsonDocument {
+}
