+++ v2/elasticsearch/src/main/java/com/google/cloud/teleport/v2/elasticsearch/transforms/PubsubMessageToFailsafeElementFn.java
@@ -1,0 +1,2 @@
+package com.google.cloud.teleport.v2.elasticsearch.transforms;public class PubsubMessageToFailsafeElementFn {
+}
