+++ v2/elasticsearch/src/main/resources/elasticsearch-command-spec.json
