--- v2/elasticsearch/src/main/resources/pubsub-to-elasticsearch-command-spec.json
@@ -1,7 +1,0 @@
-{
-  "mainClass": "com.google.cloud.teleport.v2.templates.PubSubToElasticsearch",
-  "classPath": "/template/pubsub-to-elasticsearch/*:/template/pubsub-to-elasticsearch/libs/*:/template/pubsub-to-elasticsearch/classes",
-  "defaultParameterValues": {
-    "labels": "{\"goog-dataflow-provided-template-type\":\"flex\", \"goog-dataflow-provided-template-name\":\"pubsub_to_elasticsearch\"}"
-}
-}
